requirements.txt
altgraph==0.17
asgiref==3.2.10
click==7.1.2
Django==3.1.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
django-solo==1.1.5
djangorestframework==3.12.1
djangorestframework-datatables==0.5.2
itsdangerous==1.1.0
macholib==1.14
MarkupSafe==1.1.1
numpy
pandas~=1.3.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
xlrd==1.2.0
fontawesome-free
django_celery_beat==2.2.1
django_celery_results==2.2.0
pyocclient==0.6

celery~=5.2.3